userbot/requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
fastapi==0.115.6
uvicorn==0.34.0
asyncpg==0.30.0
pydantic==2.9.2
python-dotenv==1.0.1
